+++ NEWS.md
@@ -1,4 +1,4 @@
-# shorts 1.1.2.9001 (Development version)
+# shorts 1.1.3
 
 * Added sample radar gun data from Jean-Benoît Morin Microsoft Excel spreadsheet, freely available at his [website](https://jbmorin.net/2017/12/13/a-spreadsheet-for-sprint-acceleration-force-velocity-power-profiling/) (accessed October 27, 2020)). For more infor `?jb_morin`